requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
aquacrop==2.0.4
numpy==1.21.6
scipy~=1.8.1
pathlib~=1.0.1
Flask~=2.1.2
Werkzeug~=2.1.2
SQLAlchemy~=1.4.37
psycopg2
flask
flask-cors
matplotlib
eto